requirements.txt
click>=7.0
coloredlogs>=10.0
hiredis>=1.0.1
python-rapidjson>=0.9.1
tabulate>=0.8.5
honcho==1.0.1
